feat: move `NexusApplicationCommand` to `commons`
