--- src/main/java/pw/mihou/nexus/core/managers/facade/NexusApplicationCommand.kt
@@ -1,5 +1,0 @@
-package pw.mihou.nexus.core.managers.facade
-
-interface NexusApplicationCommand {
-    val  uuid: String
-}
