+++ src/main/java/pw/mihou/nexus/features/command/facade/NexusCommand.java
@@ -5,7 +5,7 @@ import org.javacord.api.entity.permission.PermissionType;
 import org.javacord.api.interaction.*;
 import pw.mihou.nexus.core.exceptions.NoSuchAfterwareException;
 import pw.mihou.nexus.core.exceptions.NoSuchMiddlewareException;
-import pw.mihou.nexus.core.managers.facade.NexusApplicationCommand;
+import pw.mihou.nexus.features.commons.NexusApplicationCommand;
 import pw.mihou.nexus.features.command.interceptors.core.NexusCommandInterceptorCore;
 import pw.mihou.nexus.features.command.validation.OptionValidation;
