+++ src/main/java/pw/mihou/nexus/features/commons/NexusApplicationCommand.kt
@@ -1,0 +1,5 @@
+package pw.mihou.nexus.features.commons
+
+interface NexusApplicationCommand {
+    val uuid: String
+}
